tp3-commit anterior errado:aqui sim volta ao ponto antes de unsigned int

+++ tps/tp3/bib-qse-mais-tp3/area.c
@@ -96,7 +96,7 @@ void RetiraPrimeiro(TipoArea *Area, TipoRegistro *Item, int tamArea)
   Area->CelulasDisp = Area->Primeiro;
   Area->Primeiro = ProxTmp;
 
-  if ((unsigned int)Area->Primeiro < (unsigned int) tamArea)
+  if ((unsigned int)Area->Primeiro < tamArea)
     Area->Itens[Area->Primeiro].Ant = -1;
 
   Area->NumCelOcupadas--;
